% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_excel.R
\name{bulk_excel}
\alias{bulk_excel}
\title{Creating a single data frame from a folder of similarly structured Excel spreadsheets}
\usage{
bulk_excel(folder = "DEFAULTBULKCSV2017",
  export = "filenamedefaultbulkcsv2017.csv")
}
\arguments{
\item{folder}{The folder in relation to your working directory where the Excel files exist. Default folder is the current working directory.}

\item{export}{File name to export csv file as, if wanted.}
}
\description{
Like Voltron-ing your Excel data
}
\examples{
bulk_csv(folder="data", export="combined_data.csv")
}
\keyword{appending}
\keyword{data}

