% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_file.R
\name{dl_file}
\alias{dl_file}
\title{Downloads file from URL if file doesn't exist in set folder}
\usage{
dl_file(folder = "data", link = "filenamedefaultbulkcsv2017.csv")
}
\arguments{
\item{folder}{The folder in relation to your working directory where the file should exist. Default folder is `data`.}

\item{link}{The link to where the file exists}
}
\description{
Downloads file from URL if file doesn't exist in set folder
}
\examples{
dl_file(folder="data", link="https://website.com/data/bostonpayroll2013.csv")
}
\keyword{and}
\keyword{checking}
\keyword{downloading}
\keyword{file}
