% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagr.R
\name{packagr}
\alias{packagr}
\title{Downloading packages that are needed but aren't yet on the system}
\usage{
packagr(packages = c("tidyverse"), gh_packages = c("abtran/muckrakr"))
}
\arguments{
\item{packages}{Array of packages found on CRAN}

\item{gh_packages}{Array of packages found on Github}
}
\description{
Downloading packages that are needed but aren't yet on the system
}
\examples{
packagr(packages=c("tidyverse", "rtweet"), gh_packages=c("abtran/muckrakr"))
}
\keyword{Downloading}
\keyword{are}
\keyword{aren't}
\keyword{but}
\keyword{needed}
\keyword{on}
\keyword{packages}
\keyword{system}
\keyword{that}
\keyword{the}
\keyword{yet}
