% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_cap.R
\name{tweet_cap}
\alias{tweet_cap}
\title{Take a screenshot of a tweet}
\usage{
tweet_cap(link = "", filename = "NOTHINGTWEET_CAP",
  folder = "DEFAULTTWEETCAP")
}
\arguments{
\item{link}{the URL of the tweet}

\item{filename}{the base name of the image you want to save. Default will be user's twitter handle.}

\item{folder}{the folder you want to save the screenshot to}
}
\description{
Quick screenshot of a tweet based on a link
}
\examples{
tweet_cap(link="https://twitter.com/memeprovider/status/833888807959289856", folder="images")
}
\keyword{screenshot}
\keyword{twitter}

