% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untangle.R
\name{untangle}
\alias{untangle}
\title{Disentangling a complex variable}
\usage{
untangle(data, x, pattern, verbose = FALSE)
}
\arguments{
\item{data}{The name of the dataframe you}

\item{x}{Column to untangle}

\item{pattern}{Special characters that separate the variables in the column}

\item{verbose}{TRUE or FALSE}
}
\description{
We are occassionally faced with a data set variable that contains multiple pieces of information. Multiple pieces of information are stored in this variable. In this case these pieces are delineated by a comma or colon. As is, the variable is not useful.
The *untangle* function can be used to create a set of dummy codes from this variable that will be more useful. This results in a new dataset with dummy codes
}
\examples{
untangle(data=test_data, x="charges", pattern="[,:]", verbose=TRUE)
}
\keyword{data}
\keyword{wrangling}

