% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_csv.R
\name{bulk_csv}
\alias{bulk_csv}
\title{Creating a single data frame from a folder of similarly structure CSV spreadsheets}
\usage{
bulk_csv(folder = "DEFAULTBULKCSV2017",
  export = "filenamedefaultbulkcsv2017.csv")
}
\arguments{
\item{folder}{The folder in relation to your working directory where the csv files exist. Default folder is the current working directory.}

\item{export}{File name to export csv file as, if wanted.}
}
\description{
Like Voltron-ing your data
}
\examples{
bulk_csv(folder="data", export="combined_data.csv")
}
\keyword{appending}
\keyword{data}

