% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_excel.R
\name{bulk_excel}
\alias{bulk_excel}
\title{Creating a single data frame from a folder of similarly structured Excel spreadsheets}
\usage{
bulk_excel(folder = "DEFAULTBULKCSV2017",
  export = "filenamedefaultbulkcsv2017.csv", sheet = 1, col_names = TRUE,
  col_types = NULL, na = "", skip = 0)
}
\arguments{
\item{folder}{The folder in relation to your working directory where the Excel files exist. Default folder is the current working directory.}

\item{export}{File name to export csv file as, if wanted.}
}
\description{
Like Voltron-ing your Excel data. This package is based on readxl, so passing on variables from read_excel will work in bulk_excel. These variables will apply to the import of each Excel spreadsheet.
}
\examples{
bulk_excel(folder="data", export="combined_data.xls")
}
\keyword{appending}
\keyword{data}
